Corrected some light bugs

+++ include/midi.h
@@ -12,6 +12,6 @@
  * Change the cutoff, detune and amplification when the assigned knobs are being triggered
  */
 int get_midi(snd_rawmidi_t *midi_in, synth_t *synth,
-             double *attack, double *decay, double *sustain, double *release);
+             float *attack, float *decay, float *sustain, float *release);
 
 #endif
